+++ plat/kvm/memory.c
@@ -1,0 +1,10 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+/* Copyright (c) 2022, Unikraft GmbH and The Unikraft Authors.
+ * Licensed under the BSD-3-Clause License (the "License").
+ * You may not use this file except in compliance with the License.
+ */
+
+int _ukplat_mem_mappings_init(void)
+{
+	return 0;
+}
